# docs/swagger.yaml
basePath: /
definitions:
  internal_handlers.createOrderRequest:
    properties:
      price:
        example: 1200.5
        type: number
      product:
        example: Laptop
        type: string
      quantity:
        example: 1
        type: integer
    type: object
  internal_handlers.createUserRequest:
    properties:
      age:
        example: 30
        type: integer
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: securepassword
        type: string
    type: object
  internal_handlers.listUsersResponse:
    properties:
      limit:
        example: 5
        type: integer
      page:
        example: 2
        type: integer
      total:
        example: 15
        type: integer
      users:
        items:
          $ref: '#/definitions/practice_1_internal_models.User'
        type: array
    type: object
  internal_handlers.loginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: securepassword
        type: string
    type: object
  internal_handlers.loginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_handlers.updateUserRequest:
    properties:
      age:
        example: 31
        type: integer
      email:
        example: john.updated@example.com
        type: string
      name:
        example: John Updated
        type: string
    type: object
  practice_1_internal_models.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      product:
        type: string
      quantity:
        type: integer
      user_id:
        type: integer
    type: object
  practice_1_internal_models.User:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  practice_1_internal_utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing users and their orders with JWT authentication.
  title: Users and Orders API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive a JWT token
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      summary: Login
      tags:
      - auth
  /users:
    get:
      description: Get a paginated and filtered list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Minimum age
        in: query
        name: min_age
        type: integer
      - description: Maximum age
        in: query
        name: max_age
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.listUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/practice_1_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Retrieve a single user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/practice_1_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/practice_1_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{user_id}/orders:
    get:
      description: Get all orders for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/practice_1_internal_models.Order'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create an order for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Order data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/practice_1_internal_models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/practice_1_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"